[INTERNAL] MelonStorage: combine other storages for read/save

+++ melon-core/src/main/java/net/seesharpsoft/melon/StorageAdapter.java
@@ -6,14 +6,14 @@ import java.io.IOException;
 
 public interface StorageAdapter extends Comparable {
 
-    public static final int PRIORITY_LOWEST = Integer.MIN_VALUE;
+    public static final int PRIORITY_LOWEST = Byte.MIN_VALUE;
     public static final int PRIORITY_DEFAULT = 0;
-    public static final int PRIORITY_HIGHEST = Integer.MAX_VALUE;
-    
+    public static final int PRIORITY_HIGHEST = Byte.MAX_VALUE;
+
     boolean canHandle(Table table, Properties properties, Object input);
-    
+
     Storage createStorage(Table table, Properties properties, Object input) throws IOException;
-    
+
     default int getPriority() {
         return PRIORITY_DEFAULT;
     }
@@ -25,5 +25,5 @@ public interface StorageAdapter extends Comparable {
         int otherPriority = ((StorageAdapter)other).getPriority();
         return this.getPriority() > otherPriority ? -1 : this.getPriority() < otherPriority ? 1 : 0;
     }
-        
+
 }
